moved mixin to util
